moved converter control strategies to demand folder
